Fixed the problems with `gdal_merge`.
It turns out that this took a long time for me to fix because `gdal_merge` is weird about overwriting files. It does NOT overwrite, but it does not throw an error (at least not one that's caught in a jupyter notebook) when you try. Super annoyingly, it does touch the file though. So the modification date changes (and makes me think it was overwritten). The only reason I can think of to set it up this way is to play a cruel joke on anyone who wants to use it. Hilarious.